+++ app/src/main/java/com/techcourse/dispatcherServlet/DispatcherServletInitializer.java
@@ -1,6 +1,6 @@
-package com.techcourse;
+package com.techcourse.dispatcherServlet;
 
-import com.interface21.webmvc.servlet.DispatcherServlet;
+import com.interface21.webmvc.servlet.mvc.tobe.DispatcherServlet;
 import jakarta.servlet.ServletContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
