+++ app/src/main/java/com/techcourse/dispatcherServlet/handlerAdapter/ManualHandlerAdapter.java
@@ -1,6 +1,6 @@
-package com.techcourse;
+package com.techcourse.dispatcherServlet.handlerAdapter;
 
-import com.interface21.webmvc.servlet.HandlerAdapter;
+import com.interface21.webmvc.servlet.mvc.tobe.handlerAdapter.HandlerAdapter;
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.view.JspView;
