+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/handlerAdapter/HandlerAdapter.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet;
+package com.interface21.webmvc.servlet.mvc.tobe.handlerAdapter;
 
 import com.interface21.webmvc.servlet.ModelAndView;
 import jakarta.servlet.http.HttpServletRequest;
