+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/handlerMapping/HandlerMapping.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet;
+package com.interface21.webmvc.servlet.mvc.tobe.handlerMapping;
 
 import jakarta.servlet.http.HttpServletRequest;
